Move SecurityMode from leshan-server-core to leshan-core
