+++ leshan-core/src/main/java/org/eclipse/leshan/SecurityMode.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.server.bootstrap;
+package org.eclipse.leshan;
 
 /**
  * The different DTLS security modes
